+++ ToolKit/ToolKit.vcxproj.filters
@@ -156,6 +156,9 @@
     <ClCompile Include="RenderSystem.cpp">
       <Filter>Source\Render</Filter>
     </ClCompile>
+    <ClCompile Include="GBufferPass.cpp">
+      <Filter>Source\Render\Pass</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="DebugNew.h">
@@ -307,6 +310,9 @@
     </ClInclude>
     <ClInclude Include="SceneRenderer.h">
       <Filter>Source\Render</Filter>
+    </ClInclude>
+    <ClInclude Include="GBufferPass.h">
+      <Filter>Source\Render\Pass</Filter>
     </ClInclude>
   </ItemGroup>
 </Project>
